refactor to move tree and ops to core, making space for new Pedals Module
