+++ src/app/mixer/palette/operation/inlet/inlet.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 import { FormControl, FormGroupDirective, NgForm } from '@angular/forms';
-import { OperationService } from '../../../provider/operation.service';
+import { OperationService } from '../../../../core/provider/operation.service';
 import { SystemsService } from '../../../../core/provider/systems.service';
-import { TreeService } from '../../../provider/tree.service';
+import { TreeService } from '../../../../core/provider/tree.service';
 import { getDraftName } from '../../../../core/model/drafts';
 import { DynamicOperation, OperationInlet,OpNode } from '../../../../core/model/datatypes';
 import { I } from '@angular/cdk/keycodes';
