+++ src/app/mixer/palette/palette.component.ts
@@ -13,7 +13,7 @@ import { Shape } from '../model/shape';
 import utilInstance from '../../core/model/util';
 import { OperationComponent } from './operation/operation.component';
 import { ConnectionComponent } from './connection/connection.component';
-import { TreeService } from '../provider/tree.service';
+import { TreeService } from '../../core/provider/tree.service';
 import { FileService } from './../../core/provider/file.service';
 import { ViewportService } from '../provider/viewport.service';
 import { NoteComponent } from './note/note.component';
