+++ src/assets/operation_support.json
@@ -1,0 +1,11 @@
+{
+
+    "rect":{
+        "weaver_dx":"weaver description",
+        "code_dx":"coder description"
+    }
+
+
+
+
+}
